# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WhatPay API",
    "description": "Plataforma de pagos peer-to-peer integrada con WhatsApp para Chile.\n\n## Autenticación\nLa mayoría de endpoints requieren un JWT Bearer token obtenido en `POST /api/v1/users/login`.\nLos endpoints admin requieren el header `x-admin-key`.\n\n## Flujo típico\n1. Registro vía bot WhatsApp (conversacional) o `POST /api/v1/users/register`\n2. `POST /api/v1/users/login` → PIN → JWT (30 min)\n3. `POST /api/v1/topup/webpay` → recarga wallet con Transbank\n4. `POST /api/v1/payments/pay` → envía pago P2P\n5. `POST /api/v1/payments/links` → crea link de cobro para comercios\n\n## Niveles KYC\n| Nivel | Límite por tx | Límite mensual |\n|-------|--------------|----------------|\n| `BASIC` | $50.000 | $200.000 |\n| `INTERMEDIATE` | $500.000 | $2.000.000 |\n| `FULL` | $2.000.000 | Sin límite |\n\n## Seguridad\n- PIN almacenado como bcrypt hash (cost 12)\n- RUT almacenado como HMAC-SHA256 (nunca en claro)\n- Bloqueo tras 3 intentos fallidos de PIN (15 min)\n- Rate limiting por IP en todos los endpoints\n- Todas las transferencias usan SELECT FOR UPDATE (sin double-spending)\n",
    "version": "0.1.0",
    "contact": {
      "name": "Pablo Cussen",
      "url": "https://cussen.cl",
      "email": "pablo@cussen.cl"
    },
    "license": {
      "name": "Portfolio Project",
      "url": "https://cussen.cl/whatpay"
    }
  },
  "servers": [
    {
...
      "description": "Producción (GCP Cloud Run — Santiago)"
...
    {
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en POST /api/v1/users/login"
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "API key de administrador"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Datos inválidos."
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/topup/webpay": {
      "post": {
        "tags": [
          "topup"
        ],
        "summary": "Iniciar recarga con Transbank WebPay",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "integer",
                    "minimum": 1000,
                    "maximum": 500000,
                    "example": 10000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL de pago",
            "content": {
              "application/json": {
                "example": {
                  "redirectUrl": "https://webpay3g.transbank.cl/...",
                  "token": "...",
                  "amount": 10000
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
...
    "/api/v1/topup/khipu": {
      "post": {
        "tags": [
          "topup"
        ],
        "summary": "Iniciar recarga con Khipu",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "integer",
                    "minimum": 1000,
                    "maximum": 500000,
                    "example": 10000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL de pago",
            "content": {
              "application/json": {
                "example": {
                  "paymentUrl": "https://khipu.com/...",
                  "paymentId": "abc",
                  "amount": 10000
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
...
    "/api/v1/splits": {
      "post": {
        "tags": [
          "splits"
        ],
        "summary": "Crear split payment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Divide la cuenta entre N+1 participantes (creador + invitados). Máx 20 participantes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "creatorName",
                  "description",
                  "totalAmount",
                  "splitMethod",
                  "participants"
                ],
                "properties": {
                  "creatorName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "totalAmount": {
                    "type": "integer",
                    "minimum": 1000
                  },
                  "splitMethod": {
                    "type": "string",
                    "enum": [
                      "equal",
                      "custom"
                    ]
                  },
                  "participants": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "phone": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Split creado",
            "content": {
              "application/json": {
                "example": {
                  "split": {
                    "id": "spl_xxx",
                    "participants": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      },
      "get": {
        "tags": [
          "splits"
        ],
        "summary": "Mis splits",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Splits",
            "content": {
              "application/json": {
                "example": {
                  "splits": [],
                  "count": 0
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/splits/{id}": {
      "get": {
        "tags": [
          "splits"
        ],
        "summary": "Detalle de split",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Split",
            "content": {
              "application/json": {
                "example": {
                  "split": {},
                  "summary": {}
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "splits"
        ],
        "summary": "Cancelar split",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelado",
            "content": {
              "application/json": {
                "example": {
                  "message": "ok"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/splits/{id}/pay": {
      "post": {
        "tags": [
          "splits"
        ],
        "summary": "Pagar parte del split",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone",
                  "transactionRef"
                ],
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "transactionRef": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pagado",
            "content": {
              "application/json": {
                "example": {
                  "split": {}
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
    "/api/v1/splits/{id}/decline": {
      "post": {
        "tags": [
          "splits"
        ],
        "summary": "Rechazar participación",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone"
                ],
                "properties": {
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rechazado",
            "content": {
              "application/json": {
                "example": {
                  "message": "ok"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado"
          }
        }
      }
    },
...
  }
}